refactor: JsonDecoder 기능분리
