+++ Diary/Collection+.swift
@@ -5,8 +5,6 @@
 //  Created by Kiseok on 1/5/24.
 //
 
-import Foundation
-
 extension Collection {
     subscript (safe index: Index) -> Element? {
         return indices.contains(index) ? self[index] : nil
